board List 첨부파일, 답글, 답변 기능 완료

+++ src/main/java/com/community/community/notice/service/NoticeService.java
@@ -12,6 +12,6 @@ public interface NoticeService {
 	List<NoticeResponseDto> findAll(NoticeRequestDto noticeRequestDto);
 	int getTotalCount(NoticeRequestDto noticeRequestDto);
 	NoticeResponseDto findByNoticeId(Long noticeId);
-	int save(NoticeRequestDto noticeRequestDto, List<MultipartFile> files);
+	long save(NoticeRequestDto noticeRequestDto, List<MultipartFile> files);
 	int updateView(Long noticeId);
 }
